Reorganize and remove components.
