+++ assets/js/modules/adsense/components/setup/SetupAccountSite/RequiresReview.js
@@ -1,5 +1,5 @@
 /**
- * AdSense Setup Account Site Needs Attention component.
+ * AdSense Setup Account Site Requires Review component.
  *
  * Site Kit by Google, Copyright 2022 Google LLC
  *
@@ -26,13 +26,13 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import ViewContextContext from '../../../../../../components/Root/ViewContextContext';
-import { trackEvent } from '../../../../../../util';
-import { MODULES_ADSENSE } from '../../../../datastore/constants';
+import ViewContextContext from '../../../../../components/Root/ViewContextContext';
+import { trackEvent } from '../../../../../util';
+import { MODULES_ADSENSE } from '../../../datastore/constants';
 import SetupAccountSiteUI from '../common/SetupAccountSiteUI';
 const { useSelect } = Data;
 
-export default function NeedsAttention() {
+export default function RequiresReview() {
 	const viewContext = useContext( ViewContextContext );
 
 	const reviewSiteURL = useSelect( ( select ) =>
@@ -45,25 +45,22 @@ export default function NeedsAttention() {
 			trackEvent(
 				`${ viewContext }_adsense`,
 				'review_site_state',
-				'needs_attention'
+				'requires_review'
 			);
 			global.open( reviewSiteURL, '_blank' );
 		},
 		[ reviewSiteURL, viewContext ]
 	);
 
-	const heading = __(
-		'Your site isn’t ready to show ads yet',
-		'google-site-kit'
-	);
+	const heading = __( 'Your site requires review', 'google-site-kit' );
 
 	const description = __(
-		'You need to fix some things with this site before we can connect Site Kit to your AdSense account.',
+		'To start serving ads, your site needs to be approved first. Go to AdSense to request the review.',
 		'google-site-kit'
 	);
 
 	const primaryButton = {
-		label: __( 'Review site in AdSense', 'google-site-kit' ),
+		label: __( 'Request review in AdSense', 'google-site-kit' ),
 		href: reviewSiteURL,
 		onClick: reviewSiteHandler,
 	};
