+++ assets/js/modules/adsense/components/setup/SetupMain.js
@@ -1,7 +1,7 @@
 /**
  * AdSense Main setup component.
  *
- * Site Kit by Google, Copyright 2021 Google LLC
+ * Site Kit by Google, Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,245 +19,170 @@
 /**
  * External dependencies
  */
+import { useUpdateEffect } from 'react-use';
 import PropTypes from 'prop-types';
 
 /**
  * WordPress dependencies
  */
-import { useEffect, useState } from '@wordpress/element';
 import { _x } from '@wordpress/i18n';
+import { useCallback, useEffect, useState } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
+import { ProgressBar } from 'googlesitekit-components';
 import AdSenseIcon from '../../../../../svg/graphics/adsense.svg';
-import { MODULES_ADSENSE } from '../../datastore/constants';
-import { CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
-import { CORE_LOCATION } from '../../../../googlesitekit/datastore/location/constants';
-import {
-	ACCOUNT_STATUS_APPROVED,
-	SITE_STATUS_ADDED,
-	determineAccountID,
-	determineClientID,
-	determineAccountStatus,
-	determineSiteStatus,
-} from '../../util/status';
-import { getSetupMainViewComponent } from './utils';
+import SetupAccount from './SetupAccount';
+import SetupCreateAccount from './SetupCreateAccount';
+import SetupSelectAccount from './SetupSelectAccount';
 import { trackEvent } from '../../../../util';
-import { AdBlockerWarning } from '../common';
+import { AdBlockerWarning, ErrorNotices } from '../common';
+import {
+	BACKGROUND_SUBMIT_SUSPENDED,
+	MODULES_ADSENSE,
+} from '../../datastore/constants';
+import { CORE_USER } from '../../../../googlesitekit/datastore/user/constants';
+import { CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
+import {
+	ACCOUNT_STATUS_READY,
+	ACCOUNT_STATUS_NONE,
+	ACCOUNT_STATUS_MULTIPLE,
+} from '../../util/status';
 import useViewContext from '../../../../hooks/useViewContext';
+import { useRefocus } from '../../../../hooks/useRefocus';
+import { CORE_UI } from '../../../../googlesitekit/datastore/ui/constants';
 const { useSelect, useDispatch } = Data;
 
 export default function SetupMain( { finishSetup } ) {
 	const viewContext = useViewContext();
 	const eventCategory = `${ viewContext }_adsense`;
 
-	// Get settings.
-	const siteURL = useSelect( ( select ) =>
-		select( CORE_SITE ).getReferenceSiteURL()
+	const {
+		clearError,
+		resetAccounts,
+		resetAlerts,
+		resetClients,
+		resetSites,
+		setAccountID,
+		setAccountStatus,
+		submitChanges,
+	} = useDispatch( MODULES_ADSENSE );
+
+	const [ isAwaitingBackgroundSubmit, setIsAwaitingBackgroundSubmit ] =
+		useState( false );
+	// Submit changes for determined parameters in the background when they are valid.
+	const [ isSubmittingInBackground, setIsSubmittingInBackground ] =
+		useState( false );
+	const backgroundSubmissionSuspended = useSelect(
+		( select ) =>
+			!! select( CORE_UI ).getValue( BACKGROUND_SUBMIT_SUSPENDED )
 	);
-	const isNavigating = useSelect( ( select ) =>
-		select( CORE_LOCATION ).isNavigating()
+	const isAdBlockerActive = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).isAdBlockerActive()
 	);
-	const previousAccountID = useSelect( ( select ) =>
+	const accounts = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).getAccounts()
+	);
+	const accountID = useSelect( ( select ) =>
 		select( MODULES_ADSENSE ).getAccountID()
 	);
-	const previousClientID = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getClientID()
+	const hasAccountIDChanged = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).hasSettingChanged( 'accountID' )
 	);
-	const previousAccountStatus = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getAccountStatus()
+	const hasClientIDChanged = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).hasSettingChanged( 'clientID' )
 	);
-	const previousSiteStatus = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getSiteStatus()
-	);
-	const accountSetupComplete = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getAccountSetupComplete()
-	);
-	const siteSetupComplete = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getSiteSetupComplete()
-	);
-
-	// Check whether a change submission is in progress.
-	const isDoingSubmitChanges = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).isDoingSubmitChanges()
-	);
-
 	// Check whether settings differ from server and are valid.
 	const canSubmitChanges = useSelect( ( select ) =>
 		select( MODULES_ADSENSE ).canSubmitChanges()
 	);
 
-	// Determine account.
-	const accounts = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getAccounts()
+	const clientID = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).getClientID()
 	);
-	const accountID = determineAccountID( {
-		accounts,
-		previousAccountID,
-	} );
-
-	// Determine client.
-	const clients = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getClients( accountID )
+	const accountStatus = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).getAccountStatus()
 	);
-	const clientID = determineClientID( {
-		clients,
-		previousClientID,
-	} );
-
-	// Get additional information to determine account and site status.
-	const alerts = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getAlerts( accountID )
+	const hasAccountStatusChanged = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).hasSettingChanged( 'accountStatus' )
 	);
-	const urlChannels = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getURLChannels( accountID, clientID )
+	const siteStatus = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).getSiteStatus()
 	);
-	const urlChannelsError = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getErrorForSelector( 'getURLChannels', [
-			accountID,
-			clientID,
-		] )
-	);
-	const accountsError = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getError( 'getAccounts', [] )
-	);
-	const alertsError = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getError( 'getAlerts', [ accountID ] )
+	const hasSiteStatusChanged = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).hasSettingChanged( 'siteStatus' )
 	);
 	const hasErrors = useSelect( ( select ) =>
 		select( MODULES_ADSENSE ).hasErrors()
 	);
+	const hasResolvedAccounts = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).hasFinishedResolution( 'getAccounts' )
+	);
+	const userEmail = useSelect( ( select ) => select( CORE_USER ).getEmail() );
+	const referenceSiteURL = useSelect( ( select ) =>
+		select( CORE_SITE ).getReferenceSiteURL()
+	);
+	const existingTag = useSelect( ( select ) =>
+		select( MODULES_ADSENSE ).getExistingTag()
+	);
 
-	// Determine account and site status.
-	const accountStatus = determineAccountStatus( {
-		accounts,
-		clients,
-		alerts,
-		urlChannels,
-		accountsError,
-		alertsError,
-		urlChannelsError,
-		previousAccountID,
-		previousClientID,
-	} );
-	const siteStatus = determineSiteStatus( {
-		urlChannels,
-		siteURL,
-	} );
+	const account = accounts?.find( ( { _id } ) => _id === accountID );
 
-	const {
-		clearError,
-		setAccountID,
-		setClientID,
-		setAccountStatus,
-		setSiteStatus,
-		setAccountSetupComplete,
-		setSiteSetupComplete,
-		setUseSnippet,
-		submitChanges,
-		resetAccounts,
-		resetAlerts,
-		resetClients,
-		resetURLChannels,
-	} = useDispatch( MODULES_ADSENSE );
-
-	// Allow flagging when a background submission should happen.
-	const [ isAwaitingBackgroundSubmit, setIsAwaitingBackgroundSubmit ] =
-		useState( false );
+	useUpdateEffect( () => {
+		if (
+			( hasAccountIDChanged && accountID !== undefined ) ||
+			( hasClientIDChanged && clientID !== undefined ) ||
+			( hasAccountStatusChanged && accountStatus !== undefined ) ||
+			( hasSiteStatusChanged && siteStatus !== undefined )
+		) {
+			setIsAwaitingBackgroundSubmit( true );
+		}
+	}, [
+		accountID,
+		hasAccountIDChanged,
+		clientID,
+		hasClientIDChanged,
+		accountStatus,
+		hasAccountStatusChanged,
+		siteStatus,
+		hasSiteStatusChanged,
+	] );
 
 	// Update current account ID setting on-the-fly.
 	useEffect( () => {
-		// Don't do anything if setting has not loaded yet, if account ID
-		// cannot be determined, or if nothing has changed.
-		if (
-			undefined === previousAccountID ||
-			undefined === accountID ||
-			previousAccountID === accountID
-		) {
+		if ( ! Array.isArray( accounts ) ) {
 			return;
 		}
-		setAccountID( accountID );
-		// Set flag to await background submission.
-		setIsAwaitingBackgroundSubmit( true );
-	}, [ previousAccountID, accountID, setAccountID ] );
 
-	// Update current client ID setting on-the-fly.
+		let newAccountID;
+
+		if (
+			accounts.length === 1 &&
+			( ! accountID || accounts[ 0 ]._id !== accountID )
+		) {
+			newAccountID = accounts[ 0 ]._id;
+		} else if ( accounts.length === 0 && !! accountID ) {
+			newAccountID = '';
+		}
+
+		if ( newAccountID !== undefined ) {
+			setAccountID( newAccountID );
+			// Set flag to await background submission.
+			setIsAwaitingBackgroundSubmit( true );
+		}
+	}, [ accounts, accountID, setAccountID ] );
+
+	// Update account status on-the-fly.
 	useEffect( () => {
-		// Don't do anything if setting has not loaded yet, if client ID cannot
-		// be determined, or if nothing has changed.
-		if (
-			undefined === previousClientID ||
-			undefined === clientID ||
-			previousClientID === clientID
-		) {
-			return;
+		if ( accounts?.length === 0 ) {
+			setAccountStatus( ACCOUNT_STATUS_NONE );
+		} else if ( accounts?.length > 1 && ! accountID ) {
+			setAccountStatus( ACCOUNT_STATUS_MULTIPLE );
 		}
-		setClientID( clientID );
-		// Set flag to await background submission.
-		setIsAwaitingBackgroundSubmit( true );
-	}, [ previousClientID, clientID, setClientID ] );
+	}, [ setAccountStatus, accountID, accounts ] );
 
-	// Update account status setting on-the-fly.
-	useEffect( () => {
-		// Don't do anything if setting has not loaded yet, if account status
-		// cannot be determined, or if nothing has changed.
-		if (
-			undefined === previousAccountStatus ||
-			undefined === accountStatus ||
-			previousAccountStatus === accountStatus
-		) {
-			return;
-		}
-		// Force setup completion flags to false in case it had been set
-		// before, and enforce snippet placement until the account is
-		// approved.
-		if ( accountStatus !== ACCOUNT_STATUS_APPROVED ) {
-			setAccountSetupComplete( false );
-			setSiteSetupComplete( false );
-			setUseSnippet( true );
-		}
-		setAccountStatus( accountStatus );
-		// Set flag to await background submission.
-		setIsAwaitingBackgroundSubmit( true );
-	}, [
-		previousAccountStatus,
-		accountStatus,
-		setAccountSetupComplete,
-		setSiteSetupComplete,
-		setUseSnippet,
-		setAccountStatus,
-	] );
-
-	// Update site status setting on-the-fly.
-	useEffect( () => {
-		// Don't do anything if setting has not loaded yet, if site status
-		// cannot be determined, or if nothing has changed.
-		if (
-			undefined === previousSiteStatus ||
-			undefined === siteStatus ||
-			previousSiteStatus === siteStatus
-		) {
-			return;
-		}
-		// Force site setup completion flag to false in case it had been set before.
-		if ( siteStatus !== SITE_STATUS_ADDED ) {
-			setSiteSetupComplete( false );
-		}
-		setSiteStatus( siteStatus );
-		// Set flag to await background submission.
-		setIsAwaitingBackgroundSubmit( true );
-	}, [
-		previousSiteStatus,
-		siteStatus,
-		setSiteSetupComplete,
-		setSiteStatus,
-	] );
-
-	// Submit changes for determined parameters in the background when they are valid.
-	const [ isSubmittingInBackground, setIsSubmittingInBackground ] =
-		useState( false );
 	// If a background submission should happen and changes are valid to be
 	// submitted, do that here. This is wrapped in a separate useEffect hook
 	// and relies on isAwaitingBackgroundSubmit since the above useEffect hook
@@ -267,12 +192,13 @@ export default function SetupMain( { finishSetup } ) {
 		if (
 			! isAwaitingBackgroundSubmit ||
 			isSubmittingInBackground ||
-			! canSubmitChanges
+			! canSubmitChanges ||
+			backgroundSubmissionSuspended
 		) {
 			return;
 		}
 
-		// Set flag to false since we are gonna run the background submission
+		// Set flag to false since we are going to run the background submission
 		// right now.
 		setIsAwaitingBackgroundSubmit( false );
 
@@ -288,99 +214,71 @@ export default function SetupMain( { finishSetup } ) {
 		isSubmittingInBackground,
 		canSubmitChanges,
 		submitChanges,
+		backgroundSubmissionSuspended,
 	] );
 
-	const isAdBlockerActive = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).isAdBlockerActive()
-	);
+	const reset = useCallback( () => {
+		// Do not reset if account status has not been determined yet, or
+		// if the account is approved.
+		if (
+			undefined === accountStatus ||
+			ACCOUNT_STATUS_READY === accountStatus
+		) {
+			return;
+		}
 
-	// Reset all fetched data when user re-focuses tab.
-	useEffect( () => {
-		let timeout;
-		let idleSeconds = 0;
-		// Count seconds once user focuses elsewhere.
-		const countIdleTime = () => {
-			timeout = global.setInterval( () => {
-				idleSeconds++;
-			}, 1000 );
-		};
-		// Reset when user re-focuses after 15 seconds or more.
-		const reset = () => {
-			global.clearTimeout( timeout );
-			// Do not reset if user has been away for less than 15 seconds.
-			if ( idleSeconds < 15 ) {
-				idleSeconds = 0;
-				return;
-			}
-			idleSeconds = 0;
-			// Do not reset if account status has not been determined yet, or
-			// if the account is approved.
-			if (
-				undefined === accountStatus ||
-				ACCOUNT_STATUS_APPROVED === accountStatus
-			) {
-				return;
-			}
-
-			// Unset any potential error.
-			clearError();
-			// Reset all data to force re-fetch.
-			resetAccounts();
-			resetAlerts();
-			resetClients();
-			resetURLChannels();
-		};
-		global.addEventListener( 'focus', reset );
-		global.addEventListener( 'blur', countIdleTime );
-		return () => {
-			global.removeEventListener( 'focus', reset );
-			global.removeEventListener( 'blur', countIdleTime );
-			global.clearTimeout( timeout );
-		};
+		// Unset any potential error.
+		clearError();
+		// Reset all data to force re-fetch.
+		resetAccounts();
+		resetAlerts();
+		resetClients();
+		resetSites();
 	}, [
 		accountStatus,
 		clearError,
 		resetAccounts,
 		resetAlerts,
 		resetClients,
-		resetURLChannels,
+		resetSites,
 	] );
 
-	// Fetch existing tag right here, to ensure the progress bar is still being
-	// shown while this is being loaded. It is technically used only by child
-	// components.
-	const existingTag = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getExistingTag()
-	);
+	// Reset all fetched data when user re-focuses window.
+	useRefocus( reset, 15000 );
 
 	useEffect( () => {
 		if ( accountStatus !== undefined ) {
-			trackEvent(
-				eventCategory,
-				'receive_account_status',
-				accountStatus
-			);
+			trackEvent( eventCategory, 'receive_account_state', accountStatus );
 		}
 	}, [ eventCategory, accountStatus ] );
 
 	useEffect( () => {
 		if ( siteStatus !== undefined ) {
-			trackEvent( eventCategory, 'receive_site_status', siteStatus );
+			trackEvent( eventCategory, 'receive_site_state', siteStatus );
 		}
 	}, [ eventCategory, siteStatus ] );
 
-	const viewComponent = getSetupMainViewComponent(
-		accountStatus,
-		hasErrors,
-		existingTag,
-		isDoingSubmitChanges,
-		isSubmittingInBackground,
-		isNavigating,
-		accountSetupComplete,
-		siteStatus,
-		siteSetupComplete,
-		finishSetup
-	);
+	let viewComponent;
+
+	if (
+		! hasResolvedAccounts ||
+		accountID === undefined ||
+		userEmail === undefined ||
+		referenceSiteURL === undefined ||
+		existingTag === undefined
+	) {
+		viewComponent = <ProgressBar />;
+	} else if ( hasErrors ) {
+		viewComponent = <ErrorNotices />;
+	} else if ( ! accounts?.length ) {
+		viewComponent = <SetupCreateAccount />;
+	} else if ( ! accountID ) {
+		viewComponent = <SetupSelectAccount />;
+	} else {
+		viewComponent = (
+			<SetupAccount account={ account } finishSetup={ finishSetup } />
+		);
+	}
 
 	return (
 		<div className="googlesitekit-setup-module googlesitekit-setup-module--adsense">
@@ -405,8 +303,4 @@ export default function SetupMain( { finishSetup } ) {
 
 SetupMain.propTypes = {
 	finishSetup: PropTypes.func,
-};
-
-SetupMain.defaultProps = {
-	finishSetup: () => {},
 };
