--- assets/js/modules/adsense/components/setup/v2/index.js
@@ -1,27 +1,0 @@
-/**
- * AdSense Setup v2 components.
- *
- * Site Kit by Google, Copyright 2022 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export { default as SetupUseSnippetSwitch } from './SetupUseSnippetSwitch';
-export { default as SetupMain } from './SetupMain';
-export { default as SetupAccount } from './SetupAccount';
-export { default as SetupCreateAccount } from './SetupCreateAccount';
-export { default as SetupSelectAccount } from './SetupSelectAccount';
-export { default as SetupAccountNoClient } from './SetupAccountNoClient';
-export { default as SetupAccountCreateSite } from './SetupAccountCreateSite';
-export { default as SetupAccountPendingTasks } from './SetupAccountPendingTasks';
-export { default as SetupAccountSite } from './SetupAccountSite';
